Fix type issue

+++ matlab/c2xlib/SetCAMBasicVehicleContainerLowFrequency.m
@@ -15,7 +15,7 @@ classdef SetCAMBasicVehicleContainerLowFrequency < matlab.System & coder.Externa
         end
         
         function [] = stepImpl(obj, StationID, VehicleRole, ExteriorLights) 
-            tmpExteriorLights = uint32(ExteriorLights);
+            tmpExteriorLights = uint8(ExteriorLights);
             if coder.target('Rtw') || coder.target('Sfun') 
                 err = int32(0);
                 coder.cinclude(LibConfig.getCAMHeader());
